+++ src/prefiltering/QueryScoreLocal.cpp
@@ -7,7 +7,7 @@
 
 #include "QueryScoreLocal.h"
 
-QueryScoreLocal::QueryScoreLocal(int dbSize, unsigned short * seqLens, int k, short kmerThr, double kmerMatchProb, float zscoreThr)
+QueryScoreLocal::QueryScoreLocal(int dbSize, unsigned int *seqLens, int k, short kmerThr, double kmerMatchProb, float zscoreThr)
 : QueryScore(dbSize, seqLens, k, kmerThr, kmerMatchProb, zscoreThr)    // Call the QueryScore constructor
 {
     localResultSize = 0;
