+++ src/prefiltering/QueryTemplateMatcher.h
@@ -24,16 +24,16 @@ struct statistics_t{
 
 class QueryTemplateMatcher {
     public:
-        QueryTemplateMatcher (  BaseMatrix* m,
-                                IndexTable * indexTable,
-                                unsigned short * seqLens,
-                                short kmerThr,
-                                double kmerMatchProb,
-                                int kmerSize,
-                                int dbSize,
-                                bool aaBiasCorrection,
-                                int maxSeqLen,
-                                float zscoreThr){
+        QueryTemplateMatcher(BaseMatrix *m,
+                IndexTable *indexTable,
+                unsigned int *seqLens,
+                short kmerThr,
+                double kmerMatchProb,
+                int kmerSize,
+                int dbSize,
+                bool aaBiasCorrection,
+                int maxSeqLen,
+                float zscoreThr){
             this->m = m;
             this->indexTable = indexTable;
             this->kmerSize = kmerSize;
